+++ lib/tron.rb
@@ -8,6 +8,15 @@ require 'tron/failure' # Legacy
 
 module Tron
   def self.success(code, attributes = {}) # rubocop:disable Metrics/MethodLength
+    code.respond_to?(:to_sym) ||
+      raise(ArgumentError, 'Tron.success must be called with a Symbol as first argument')
+
+    attributes.respond_to?(:keys)||
+      raise(ArgumentError, 'The attributes Hash for Tron.success must respond to #keys')
+
+    attributes.respond_to?(:values) ||
+      raise(ArgumentError, 'The attributes Hash for Tron.success must respond to #values')
+
     Struct.new(:success, *attributes.keys) do
       undef_method '[]='.to_sym
       members.each { |member| undef_method "#{member}=".to_sym }
@@ -20,6 +29,10 @@ module Tron
         false
       end
 
+      def failure
+        nil
+      end
+
       def on_success(proc = nil, &block)
         (proc || block).call
       end
@@ -27,10 +40,19 @@ module Tron
       def on_failure(_ = nil)
         self
       end
-    end.new code.freeze, *attributes.values.map(&:freeze)
+    end.new code.to_sym, *attributes.values.map(&:freeze)
   end
 
   def self.failure(code, attributes = {}) # rubocop:disable Metrics/MethodLength
+    code.respond_to?(:to_sym) ||
+      raise(ArgumentError, 'Tron.failure must be called with a Symbol as first argument')
+
+    attributes.respond_to?(:keys)||
+      raise(ArgumentError, 'The attributes Hash for Tron.failure must respond to #keys')
+
+    attributes.respond_to?(:values) ||
+      raise(ArgumentError, 'The attributes Hash for Tron.failure must respond to #values')
+
     Struct.new(:failure, *attributes.keys) do
       undef_method '[]='.to_sym
       members.each { |member| undef_method "#{member}=".to_sym }
@@ -43,6 +65,10 @@ module Tron
         true
       end
 
+      def success
+        nil
+      end
+
       def on_success(_ = nil)
         self
       end
@@ -50,6 +76,6 @@ module Tron
       def on_failure(proc = nil, &block)
         (proc || block).call
       end
-    end.new code.freeze, *attributes.values.map(&:freeze)
+    end.new code.to_sym, *attributes.values.map(&:freeze)
   end
 end